+++ Compiler/Parser/parser.py
@@ -11,15 +11,10 @@ from Compiler.Compiletime.syntax_check import *
 from Compiler.Tokenizer.tokenizer import Token, TokenType, to_str, tokenize
 from Compiler.Parser.parserChecker import *
 from Compiler.Parser.parserTokenToNode import *
+from Compiler.Parser.parserVisits import *
+from Compiler.utils import error_list
 import os
-'''
-This variable will hold the current path of the file to be parsed.
-'''
-_curr_path = ''
-'''
-This list will be holding all the errors tracked during parsing
-'''
-error_list = []
+
 
 class Scope:
     def __init__(self, level: int, of_: NodeTypes, ended: bool) -> None:
@@ -377,48 +372,3 @@ def Compile(code: list) -> str:
     return code_string
 
 
-'''
-Function to import code on the basis of given ImportNode
-'''
-def visit_ImportNode(node):
-    csq_include_path = os.getenv("CSQ_INCLUDE")
-    importPath = os.path.join(csq_include_path, "Core", "Include", "Import")
-   
-    modulePath = None
-
-    if os.path.isfile(os.path.join(_curr_path , node.path+".csq")):
-        modulePath = os.path.join(_curr_path , node.path+".csq")
-    else:
-        modulePath = os.path.join(importPath , node.path+".csq")
-    
-    module = open(modulePath, "r")
-    # Read the file and process it
-    code_ = module.read()
-    # Convert it into stream of tokens
-    lines = []
-    for line in code_.split("\n"):
-        if line != "":
-            lines.append(tokenize(line))
-
-    # Moving forth to compilation
-    compiled_code = Compile(lines)
-    return compiled_code
-'''
-Function to import C/C++ code on the basis of given CImportNode
-'''
-def visit_CImportNode(node):
-    csq_include_path = os.getenv("CSQ_INCLUDE")
-    cimportPath = os.path.join(csq_include_path, "Core", "Include", "Cimport")
-   
-    modulePath = None
-
-    if os.path.isfile(os.path.join(_curr_path , node.path+".cpp")):
-        modulePath = os.path.join(_curr_path , node.path+".cpp")
-    else:
-        modulePath = os.path.join(cimportPath , node.path+".cpp")
-    
-    module = open(modulePath, "r")
-
-    # Read the file and process it
-    code_ = module.read()
-    return code_
