+++ Compiler/utils.py
@@ -1,0 +1,8 @@
+'''
+This variable will hold the current path of the file to be parsed.
+'''
+_curr_path = ''
+'''
+This list will be holding all the errors tracked during parsing
+'''
+error_list = []
